Allow inclusion of all subjects
Adjust script to allow for inclusion of all subjects.
The script will check whether the desired files exist. If not, warning is raised. If exist, appropriate action is taken.

+++ behavioral/REV_SST/scripts/analyses/prep4analysis.m
@@ -5,9 +5,10 @@
 
 % INITIAL SETUP (may need to change these parameters)
 %cd('~/Desktop/REV_scripts/behavioral/REV_SST/output/')
-cd('~/Desktop/REV/REV_scripts/behavioral/REV_SST/output/')
+cd('~/Desktop/REV_scripts/behavioral/REV_SST/output/')
 prefix = 'REV';
-Subs=[1:3 6 9 10:11 13 16:24 26:27 29 31 32 34:39 41 43 44 46:60 62 65 67:70 73:78 80 82 84 86 88:91 93 94 97 98 100 102 104 107:111 114:119 121 124 126:127 129:131 134:138 140:142 144];
+Subs=[1:144]
+    %1:3 6 9 10:11 13 16:24 26:27 29 31 32 34:39 41 43 44 46:60 62 65 67:70 73:78 80 82 84 86 88:91 93 94 97 98 100 102 104 107:111 114:119 121 124 126:127 129:131 134:138 140:142 144];
 %startSub = 1;
 %endSub = 144;
 startRun = 1;
@@ -15,17 +16,42 @@ endRun = 14;
 
 % COPY AND RENAME FILES
 for s=Subs   %for s=startSub:endSub
-    cd pre
-    copyfile(['sub' num2str(s) '*.mat'],'../analysisReady')
-    cd ..
+ 
+    if exist('pre', 'dir')
+        cd pre
+        try
+            copyfile(['sub' num2str(s) '*.mat'],'../analysisReady')
+        catch
+            warning('pre data for %d does not exist',s)
+        end
+        cd ..
+    else 
+        warning('pre dir does not exist')
+    end
     
-    cd post
-    copyfile(['sub' num2str(s) '*.mat'],'../analysisReady')
-    cd ..
+    if exist('post', 'dir')
+        cd post
+        try
+            copyfile(['sub' num2str(s) '*.mat'],'../analysisReady')
+        catch
+            warning('post data for %d does not exist',s)
+        end
+        cd ..
+    else 
+        warning('post dir does not exist')
+    end    
     
-    cd train
-    copyfile(['sub' num2str(s) '*.mat'],'../analysisReady')
-    cd ..
+    if exist('train', 'dir')
+        cd train
+        try
+            copyfile(['sub' num2str(s) '*.mat'],'../analysisReady')
+        catch
+            warning('training data for %d does not exist',s)
+        end
+        cd ..
+    else 
+        warning('training dir does not exist')
+    end 
     
     cd analysisReady
